+++ packages/thirdweb/src/exports/wallets.native.ts
@@ -3,22 +3,22 @@
 export {
   createWallet,
   walletConnect,
-} from "../../wallets/native/create-wallet.js";
-export { inAppWallet } from "../../wallets/in-app/native/in-app.js";
-export { smartWallet } from "../../wallets/smart/smart-wallet.js";
+} from "../wallets/native/create-wallet.js";
+export { inAppWallet } from "../wallets/in-app/native/in-app.js";
+export { smartWallet } from "../wallets/smart/smart-wallet.js";
 
-export type { Wallet, Account } from "../../wallets/interfaces/wallet.js";
-export type { ConnectionStatus } from "../../wallets/manager/index.js";
+export type { Wallet, Account } from "../wallets/interfaces/wallet.js";
+export type { ConnectionStatus } from "../wallets/manager/index.js";
 
 // utils
 export {
   getWalletBalance,
   type GetWalletBalanceOptions,
-} from "../../wallets/utils/getWalletBalance.js";
+} from "../wallets/utils/getWalletBalance.js";
 export {
   generateAccount,
   type GenerateAccountOptions,
-} from "../../wallets/utils/generateAccount.js";
+} from "../wallets/utils/generateAccount.js";
 
 // private-key
 export {
@@ -34,7 +34,7 @@ export {
    * @deprecated - use {@link PrivateKeyToAccountOptions} instead
    */
   type PrivateKeyToAccountOptions as PrivateKeyAccountOptions,
-} from "../../wallets/private-key.js";
+} from "../wallets/private-key.js";
 
 export type {
   WalletId,
@@ -42,22 +42,22 @@ export type {
   WalletCreationOptions,
   WalletConnectionOption,
   CreateWalletArgs,
-} from "../../wallets/wallet-types.js";
+} from "../wallets/wallet-types.js";
 
 export type {
   WCSupportedWalletIds,
   InjectedSupportedWalletIds,
-} from "../../wallets/__generated__/wallet-ids.js";
+} from "../wallets/__generated__/wallet-ids.js";
 
 export type {
   WCConnectOptions,
   WCAutoConnectOptions,
-} from "../../wallets/wallet-connect/types.js";
+} from "../wallets/wallet-connect/types.js";
 
 export type {
   SmartWalletConnectionOptions,
   SmartWalletOptions,
-} from "../../wallets/smart/types.js";
+} from "../wallets/smart/types.js";
 
 export type {
   InAppWalletAuth,
@@ -80,20 +80,20 @@ export type {
    * @deprecated use InAppWalletSocialAuth instead
    */
   InAppWalletSocialAuth as EmbeddedWalletSocialAuth,
-} from "../../wallets/in-app/core/wallet/types.js";
+} from "../wallets/in-app/core/wallet/types.js";
 
-export type { CoinbaseSDKWalletConnectionOptions } from "../../wallets/coinbase/coinbaseSDKWallet.js";
+export type { CoinbaseSDKWalletConnectionOptions } from "../wallets/coinbase/coinbaseSDKWallet.js";
 
 export type {
   WalletEmitter,
   WalletEmitterEvents,
-} from "../../wallets/wallet-emitter.js";
+} from "../wallets/wallet-emitter.js";
 
-export { getAllWalletsList } from "../../wallets/getAllWalletsList.js";
-export { getWalletInfo } from "../../wallets/__generated__/getWalletInfo.js";
-export { type WalletInfo } from "../../wallets/wallet-info.js";
+export { getAllWalletsList } from "../wallets/getAllWalletsList.js";
+export { getWalletInfo } from "../wallets/__generated__/getWalletInfo.js";
+export { type WalletInfo } from "../wallets/wallet-info.js";
 
-export { createWalletAdapter } from "../../adapters/wallet-adapter.js";
+export { createWalletAdapter } from "../adapters/wallet-adapter.js";
 
 // NOT SUPPORTED
